# Conflicts:
#	src/Workspaces/VisualBasic/Portable/Simplification/Reducers/VisualBasicInferredMemberNameReducer.vb

+++ src/Compilers/Core/CommandLine/NativeMethods.cs
@@ -50,7 +50,7 @@ namespace Microsoft.CodeAnalysis.CommandLine
         #region Constants
 
         internal static readonly IntPtr NullPtr = IntPtr.Zero;
-        internal static readonly IntPtr InvalidIntPtr = new IntPtr((int)-1);
+        internal static readonly IntPtr InvalidIntPtr = new IntPtr(-1);
 
         internal const uint NORMAL_PRIORITY_CLASS = 0x0020;
         internal const uint CREATE_NO_WINDOW = 0x08000000;
@@ -74,7 +74,7 @@ namespace Microsoft.CodeAnalysis.CommandLine
         internal static extern bool CreateProcess
         (
             string? lpApplicationName,
-            [In, Out]StringBuilder lpCommandLine,
+            [In, Out] StringBuilder lpCommandLine,
             IntPtr lpProcessAttributes,
             IntPtr lpThreadAttributes,
             [In, MarshalAs(UnmanagedType.Bool)]
